refactor: sort component servers into dedicated files
